+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- n/a
+### Added
+- Added a new attribute a`relative_url` to `Distributor`, so users can search distributors
+  by relative_url
 
 ## [2.1.0] - 2019-09-10